chore: refactored file structure 2/2
