+++ app/components/Views/TrendingView/Views/ExploreSearchScreen/ExploreSearchScreen.tsx
@@ -3,10 +3,10 @@ import { Keyboard } from 'react-native';
 import { useSafeAreaInsets } from 'react-native-safe-area-context';
 import { useNavigation } from '@react-navigation/native';
 import { Box } from '@metamask/design-system-react-native';
-import ExploreSearchBar from '../ExploreSearchBar/ExploreSearchBar';
-import ExploreSearchResults from './components/ExploreSearchResults/ExploreSearchResults';
-import { PerpsStreamProvider } from '../../../UI/Perps/providers/PerpsStreamManager';
-import { PerpsConnectionProvider } from '../../../UI/Perps/providers/PerpsConnectionProvider';
+import ExploreSearchBar from '../../components/ExploreSearchBar/ExploreSearchBar';
+import ExploreSearchResults from '../../components/ExploreSearchResults/ExploreSearchResults';
+import { PerpsStreamProvider } from '../../../../UI/Perps/providers/PerpsStreamManager';
+import { PerpsConnectionProvider } from '../../../../UI/Perps/providers/PerpsConnectionProvider';
 
 const ExploreSearchScreen: React.FC = () => {
   const insets = useSafeAreaInsets();
